refactor: rename QuotePolicy→SimplePolicy, split policy into submodules
- Split monolithic src/policy.rs into policy/{mod,simple,rego}.rs
- Rename QuotePolicy to SimplePolicy for clarity
- Advisory skip only during platform grace for pure OutOfDate
  (collateral grace and OutOfDateConfigurationNeeded don't skip)
- Update all imports and re-exports across lib, cli, verify
